// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  http: typeof http;
  notes: typeof notes;
  user: typeof user;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;